less-confusing regret values when deviation data is missing
Previously, when profiles were missing, regret() would perform the calculation
only over the available profiles (on the assumption that this is slightly more
informative than refusing to calculate anything). However, this caused confusing
regret values when regret due to some deviations could be calculated, but others
couldn't. Users seem to prefer seeing only the known regret values, not the
partial-data estimeates, so I've updated so that both pure and mixed regret
calculations return -inf when required profiles are missing.