Fix behavior change in subgame
add strategy used to silently work if strategy was already in the
subgame. A recent change caused it to fail with a somewhat arcane
assertion error if strategy was already in the subgame. This change
reverts that behavior, so now adding a strategy that already exists will
silently work and return the same subgame.