Modified optimization nash method
Before nash optimization enforced that a point is on the simplex with
penalty terms, now that's only done for the positivity constraint. The
point is ensured to sum to one by projecting the gradient such that it
is always constant sum. This will only work if the initial point is
feasible (or at least sums to 1).