Add max_mix_social_welfare to regret
Also additional changes to support this, such as a jocobian for expected
value, and an is_complete() for games that define values for all
profiles.

There's also a bug fix for simplex_project which failed when elements
were significantly large.

That referenced function computes the maximum role symmetric mixed
social welfare of a complete game. It's not guaranteed to find the
global optimum since the function for mixed social welfare isn't convex,
but it tries a bunch of starting points and can be given random restarts
too.